+++ lua/plugins/setup.lua
@@ -1,18 +1,18 @@
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+local lazypath = vim.fn.stdpath('data') .. '/lazy/lazy.nvim'
 
 if not (vim.uv or vim.loop).fs_stat(lazypath) then
 	vim.fn.system({
-		"git",
-		"clone",
-		"--filter=blob:none",
-		"https://github.com/folke/lazy.nvim.git",
-		"--branch=stable",
+		'git',
+		'clone',
+		'--filter=blob:none',
+		'https://github.com/folke/lazy.nvim.git',
+		'--branch=stable',
 		lazypath,
 	})
 end
 
 vim.opt.rtp:prepend(lazypath)
 
-local plugins = require("plugins.plugins")
+local plugins = require 'plugins.install'
 
-require("lazy").setup(plugins)
+require('lazy').setup(plugins)
